MDEV-19212: large page allocations over 4G overflow
When introduced, the type of os_large_page_size was dropped
from ulint to uint. The assumption is this because the
opt_large_page_size size was uint. The consequence was the
explicit type of the memory allocation was smaller resulting
in an integer overflow. Less memory is allocated, less than
what was specified.

Fixes: 2e814d4702d71a04388386a9f591d14a35980bfe

ref: MySQL Bug #94747 - "4GB Limit on large_pages shared memory set-up"
thanks to Nikolai Ikhalainen.