mysql_install_db: assume builddir is dirname0
The assumption in the original commit for --builddir (648d3cedbc09),
was to assume that the without a --builddir, and when --srcdir
is specified, that the builddir is the same as the srcdir.

As mysql_install_db is in the builddir, we use dirname0 after a quick
check that my_print_defaults is also located from dirname0.