After-merge fix: Correct a copyright statement
In the merge c51f85f8823a845cd4a6aa1b2aa5af18484b2ab0
the files row0trunc.cc and read0read.cc were not supposed
to be changed in MariaDB 10.3.

The row0trunc.cc code that was changed in MariaDB 10.2
commit 3db94d2403c1f2791bddd43656d7b0d6320b453d does not exist,
because it was removed earlier in 10.3
commit 5a1868b58d26b286b6ad433096e7184895953311.

When it comes to read0read.cc, as explained in the 10.2
commit e0271a7b43c6df652c6a074858853a6d0da20c1e
the code had been moved to trx0sys.h in 10.3. In
10.3 commit 380069c235cf63cedff1855314766a545fb26846
appproriate mutex protection was added to checking the
transaction system state. The assertion expression was
properly updated in the earlier merge
commit acf6f92aa936fbfe7524617ae57d011ab8f1f96d.